requirements.txt
boto3==1.21.22
botocore==1.24.22
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.0.4
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==1.0.0
MarkupSafe==2.1.1
Pillow==9.0.1
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.3
